Add rule for requiring HTTP 415 on all resources
As noted in #35, we should require that these status codes are present,
as it makes it clearer to the consumer that they are expected.

Although producers of APIs may not think they are relevant, as they
could be handled by the web framework, we should make it clear that this
is still required.