Add new rules about metadata
Description and contact are useful for making APIs more discoverable, and in
particular for generating catalogues.

I've updated some of the other tests that started failing because they were
missing description and contact fields to use `resultsForCode` instead of
`getErrors`, as this allows us to test a specific thing rather than having to
update all the test files every time we add a new rule. The exception is
`complete.test.js`, which is testing that everything that is required is
present.

For the contact test it's unfortunately not possible to test that the `contact`
object has either a url or email present, which would be ideal, so this just
tests that there is a contact object present.