Update the domain rules
Split this into 2 rules, one to check that the domain is the correct length and
one to check it contains the correct characters. This makes it easier to see
why this causes an error if the domain doesn't follow one or both of these
rules.

We also need to include the scheme in the regex, which should be `https` unless
the url is `http://localhost`.