Add the strictest ruleset
This contains the most opinionated rules of the three rulesets.  A couple of
rules (unsupported media type and not acceptable) have been moved into this one
as these aren't always essential and should be in the stricter ruleset.

Some of the rules add to the ones in the medium ruleset, for example the
license rule and oas version rule are more specific.